services/graph-builder/requirements.txt
kafka-python
neo4j
